ai_jarvis_assistant/requirements.txt
# Core AI assistant requirements
torch==2.2.2
transformers==4.37.2
TTS==0.22.0
openai-whisper
faster-whisper

# Audio input/output
PyAudioWPatch==0.2.12.7
pvporcupine==2.2.1
pydub==0.25.1
soundfile==0.12.1
speechrecognition==3.10.1

# Supporting dependencies
numpy==1.26.4
numba==0.58.1
llvmlite==0.41.1
scipy==1.13.0
librosa==0.10.0.post2
soxr==0.3.7
pandas==2.2.2
matplotlib==3.8.4
python-dotenv==1.0.1
requests==2.31.0
